backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
PyMuPDF==1.23.14
python-jose[cryptography]==3.3.0
scikit-learn==1.3.2
sentence-transformers>=2.3.0
huggingface_hub>=0.17.0
numpy==1.24.4
pandas==2.1.4
pyarrow==14.0.1
anthropic==0.40.0
jinja2==3.1.2
markdown==3.5.1
weasyprint==60.2
httpx==0.27.2
groq==0.29.0
faker==19.12.0
redis==5.0.1
rq==1.15.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
tenacity==8.2.3
nltk>=3.8.2